move courseName to shared folder and update references
